Update solution after mesh refinement in worker
This commit removes storing and updating the solution vector in Refiner
and delegates this to the worker. This is better, as the worker is in a
mach better position to do this update, especially, since node-, vs
cell-based schemes have to do this update very differently and thus this
does not contaminate and complicate code in Refiner.