Modify Discretization children ctor args
Instead of creating facedata and passing it to Discretization children's
ctor, we now pass the data needed to construct facedata. Though this may
not be strictly necessary for the next steps, this appears cleaner as
nodal discretization schemes are free of the dependency on FaceData,
though currently they construct it.

Also remove DG::m_vol, unused.