Add optional blocking migration
This commit adds an optional blocking migration capability to inciter.
By default blocking migration is off and we use migration in
non-blocking mode.  Blocking migration can be enabled by the -m or
--blocking command line arguments to inciter.

At this time, compared to non-blocking migration blocking migration
appears to be more error-prone: (1) it cannot be used together with
quiescence (deadlock) detection, and (2) randomly deadlocks - hence it
is not the default and no regression test exercise this.

This commit keeps the asynclogic migration tests disabled while we
resolve the problems and misuse of migration.

More background:

If a +balancer Charm++ argument is given, Charm++ turns on automatic
load balancing. So with this commit we turn that off very early on in an
initproc routine in the new class LBSwitch, defined under Main/. We then
only turn on automatic load balancing when time stepping starts. This
appears necessary for blocking migration as not all of the bound arrays
are alive during setup and thus should not be migrated.

More details at Charm++ manual, Sec. Load Balancing.