Modify DG's time stepping logic to account for returning from mesh refinement
This commit removes the dependency of the nodal discretization classes
(ALECG, and DiagCG) from FaceData. FaceData is really only used for
cell-based discretizations, e.g., DG, and not needed. FaceData's
constructor is also changed slightly by NOT doing the remapping of the
boundary-face connectivity inside FaceData's constructor but requiring
the client to feed the correct data into FaceData. This is done via a
new remap overload in Mesh/Reorder.[Ch], taking a vector of integer ids
and a hash-map.  This separates concerns and responsibilities better.

Accordingly with the above, ALECG and DiagCG now do not store a FaceData
object, but only the boundary node lists associated to side sets in
m_bnode. DG still stores a FaceData object (m_fd) but only populates the
cell-based data inside. (Note that this was already like this, because
Transporter::createPartitioner() only reads node lists if a nodal or
face listrs for cell-centered discretizations, respectively, but
explicitly passing an empty container makes this more visible.)

DG/DiagCG/ALECG::m_initial is now an integer but its job is the same: to
indicate whether (par of) the setup within these classes are run before
(m_initial=1) or during (m_initial=0) time stepping. This is used to
differentiate code paths executed during setup or reinitializing the
mesh and related data returning from mesh refinement during time
stepping.

This commit also modifies DG's time stepping logic to accommodate
returning from mesh refinement. Calling Transporter::comfinal now sums
the value of m_initial so that it can take a different action depending
on we are during setup or times stepping.

Note that at time, with these changes, all regression tests pass, but DG
does not appear to return from mesh refinement with the correct bface
and triinpoel data that is consistent with the volume elements (which
appear correct).  Also left in some debug code for now. Also, returning
from a mesh refinement step, we do not yet attempt to update the
solution on the new mesh.