Generate continuous face ids for saving side sets
This fixes a bug that sneaked in with d6e97e63, fixes based on
cppcheck's suggestions. Replacing a range-for with std::iota is cool,
but std::iota takes the initial value for its counter by value and thus
we either keep the range for or increase the counter by the number of
faces after we are done iwth a side set, so that the counter value gets
propagated to the outer scope of std::iota.