Write ordered side set face connectivity
This commit puts in code to order the side set triangle faces consistent
with the generated exodus-internal face ordering before writing them to
file.

Background: When we output a mesh with side sets consisting of boundary
face ids and their connectivity Exodus offers a choice:

(1) Write the face id pointing to a tet (a tet in a tetrahedron element
block) and write out an element-relative side id for each face on each
side set that associates which side of a tet coincides with the given
side set, or

(2) Write the face id pointing to a triangle (a triangle in a triangle
element block) and write out an element-relative side id for each face
on each side set that associates "which side of a triangle coincides
with the given side set".

When there is side sets to be written, ExodusIIMeshWriter always opts
for option (2) because the element-relative side ids for triangle faces
is always 1 so that data is easy to generate.

This commit augments this procedure by ordering the triangle
connectivity for the side sets in the order we write the face ids for
each side set to the file, so that the file-internal face id, generated
before writing the data out, points to the correct triangle connectivity
in the triangle element block.

This commit also simplifies the boundary data structure in
Refiner::bnd() and thus also how it is computed. We still store all tet
ids adjacent to all triangle elements for all physical and chare
boundaries, but we no longer store the set id, which is then searched
later instead if needed. The set id we are taking out from the container
returned by Refiner::boundary() was only used by the algorithm that
regenerated the side set faces for the refined mesh but not the
algorithm that regenerated the node lists of the refined mesh associated
to side sets. (The container returned from Refiner::boundary() is used
by both algorithms regenerating the node lists and the face lists for
each side set.)