DG now correctly returns from a mesh refinement step
Correctly regenerating the ghost data, etc.

DG now also correctly updates the solution for a refined mesh for P0,
but only after the first refinement step, not the second one. The
solution is update in DG::resize(), consistent with DiagCG.

Replace complex if condition with a bool, DG::m_refined, that indicates
whether the mesh has been refined in the current time step. This is
better, because the logic that triggers mesh refinement is stated only
once.

The new overloads to AMR is so that the functions can be called in a
const context.

Add unit tests to new remap functions in Mesh/Reorder.[Ch].