Ensure that DG soln is not read and written simultaneously
This commit ensures that after a mesh refinement step when we update the
DG solution vector, the vector is not being read and written at the same
time. Otherwise, this leads to a corrupted solution.

We also add the -m argument to a couple of single-PE DG/AMR regression
tests. This is needed because of the use of unordered containers + these
regression test baselines have been setup in non-SMP mode.