Allow parsing line with whitespaces + '\n'
This commit fixes a problem with parsing blocks in which an empty line
ONLY containing one or more withspaces would not be correctly ignored by
the parser. This affects all executables and all parsed blocks.

Example that would previously not be correctly parsed, triggering an
"unknown keyword" parser error:

   #edgelist
      #54 37 #27 53 27 55
      #37 51 37 54 37 55
      #46 26 46 39 46 44

      #39 37 39 42 39 44
   #end

Note the empty-looking line after the third list of numbers, which
contains some white spaces. This is now correctly parsed.