Refiner::writeMesh now saves elem-centered data
This commit puts in code to save the evaluated initial conditions during
initial mesh refinement using DG. Though not yet tested with regression
tests, appears to work.

We also make some changes to the AMR lib because the
tet_store::get_refinement_level_list() and get_cell_type_list() member
functions should be a const. This allows Refiner::writeMesh() to be
const and thus the compiler also selects the correct overload for
tk::remap(), to remap the triangle/face connectivty to from global to
local ids for output to file.

This also removes the need to store the cell_type_list and the
refinement_level_list in tet_store_t - which nothing else used.