Replace the internals of tagged tuple with brigand code
This commit replaces the implementation of tk::tuple::tagged_tuple<>
with tk::TaggedTuple<> using the brigand template meta-programming
library.  We have been using brigand before this change as well. While
brigand only requires C++11, the API that tk::TaggedTuple provides
requires C++14 (auto return types), and will probably soon C++17 (fold
expressions), features.

The reason for this change is to remove a huge amount of
extremely-hard-to-read code from the tuple, which previously used
had-coded type lists, compile-time indexing, and other compile-time type
manipulations, which is now hidden away in brigand, and to make the API
cleaner and easier to use correctly.

There is more work to be done to improve the client-side API for
TaggedTuple, but that is left for future work.

Some changes compared to the previous API:

* The new class requires its type list as a brigand::list. This could
have been made even simpler for client-code with a raw type list, but
requiring a brigand::list avoids having to repeat the type list twice
when declaring and defining a class inheriting from a tagged tuple that
is PUPable, which is less error-prone and easier to use for a slightly
more verbose client interface.

* The initializer constructor now takes a tuple, instead of its variadic
member list.

* There is no 'set' member function. Its functionality is replaced by
'get' returning a non-const reference.

* The intermediate API, between parsers and the tuple, in Control.h, is
now gone.