Find backtrackers's faces after mesh refinement
This commit augments the code in Refiner that updates the boundary face
data structures to detect and match faces of refined elements, adjacent
to boundaries labeled with side sets, to also detect backtrackers.

Backtrackers are tetrahedra that undergo 2->8 or 4->8 refinement, which
involve a backtracking to the single-cell, i.e., grandparent, level and
applying a 1->8 refinement. Because such backtracking involves 3 levels
of the mesh, the code that finds boundary faces associated to side sets
of children and parents needed to be extended.

This appears to work in serial only for now.