Correct communication, correction, and boundary data regeneration logic after refinement
Huge thanks to @rfbird for the help with all this.

* In Refiner, the communication of edge refinement data (tags, locks, and
intermediates) before the correction iteration is now also iterative. The send
is in Refiner::comExtra(), the receive is in Refiner::addRefBndEdges(), and the
exit condition is tested in Transporter::compatibility(). This is because in
addRefBndEdges(), when all contributions have been received by a Refiner chare,
we now run the compatibility algorithm after intermediates have been applied.
We iterate on the communication of the edge data followed by applying
compatibility until no change is detected in the refinement data on any chare.

* When the edge refinement data is agreed on across partitions, as before we
proceed to Refiner::correctref(), which performs the correction algorithm
considering the local and remote refinement data for chare-boundary edges. This
commit improves the correction logic by adding a new part of the if condition
that triggers an additional correction step: when the local edge state changes
and the remote needs to know about it. (Previously, only the case when the
local state changed but not the remote triggered an additional correction
step.)

* Also put in an assert into AMR::mesh_adapter_t::mark_error_refinement_corr()
that triggers if a lock is erroneously about to be overwritten with a weaker
lock state, e.g., lock -> unlock.

* In AMR::mesh_adapter_t::node_connectivity_t::get() changed how a node that is
not found handled. Instead of triggering an error if no parent of a node can be
found, now a pair of the input node id is returned. This is now an indication
of reaching the parent in the coarse (original) mesh. We now rely on this to
iteratively query the original parents of a node from arbitrary refinement
depths.

* During the setup of the DG communication data, instead of only counting the
expected and received chare-boundary faces, we now instead store all expected
faces and at the end of the communication step, we verify that indeed all
expected faces have indeed been received on a DG chare, no more no less and
exactly those that are expected. If this is not true, we generate a nice error
message than before and list the missing faces and their coordinates.  All of
this only happens in DEBUG mode of course.

* When Refiner regenerates the node-, and face lists for the refined mesh, we
now use algorithms that always search in the original (coarsest) mesh instead
of the previous mesh. The algorithm querying the previous mesh was lot more
complex, had to rely on a 'grandparent' mesh for backtrackers (tets that
underwent 2->8 or 4->8 refinement), and was also incorrect in some case. Such a
case is when a backtracking happens across more that 3 refinement steps: e.g.,
1, 1->2, 2, 2->8. In the third step the cell is not refined, yet in the last
step, the grandparent now should be not the second but the first mesh.
Additionally, there can be any number of steps in which the cell is not refined
(stays 2), which would necessitate storing data for all previous steps for
backtrackers. The new algorithm recursively queries only the parents in the
coarsest (original) mesh and searches its nodes and face lists. This is not
only correct, but simpler, and perhaps even more performant despite of being
recursive in the parent search. The recursive parent search is implemented by
Refiner::ancestors().

* After the boundary face lists and connectivity is updated after a refinement
step, we now always compute a surface integral across the whole problem to
verify that the newly updated data does not yield a leak. If it does, that's an
error, because of either not all or too many (e.g., multiply-counted) faces
have been found. All of this only happens in DEBUG only. The partial boundary
surface integrals are computed in Refiner::bndIntegral() and final result is
contributed to Transporter::bndint() which triggers an error if the integral is
nonzero.