Refiner now evaluates errors in mesh cells
Errors for nodal discretizations, e.g., DiagCG, are evaluated in edges
so they can be tagged for refinement or de-refinement. This is now
transferred to cells for field output.

Previously, this error was displayed as a nodal field output, but that
would require communication and more complex code to visualize the
correct errors in parallel. Transferring edge errors to cells are done
by simply averaging the 6 edges of a tet to a tet. Cell-based field
output of errors will also be useful for DG for which we will define
errors in cells.