Remove all-to-all sync between communicating the unlimited and limited solution in DG
This is done by introducing three small data structures:

 1. m_bid, a map between local and boundary ghost tet ids
 2. m_uc, solution receive buffer
 3. m_ndofc, number of degrees of freedom receive buffer

The sizes of all three equal to the number of ghost tets in a chare. The
receive buffers are the ones written during communication of the limited
solution/ndof data in DG::comlim(). Since during receiving the limited
data is written into m_uc and m_ndofc instead of m_u and m_ndof (and
only after all contributions have been received on a chare are these
merged in to m_u and m_ndof), the communication of the unlimited and
limited solution/ndof can now be overlapped and the (all-to-all) barrier
between the two communication steps is eliminated.

Note: m_bid still needs to be updated after dtref mesh refinement. This
is not yet done, because we have no tests that exercise h-ref + DG(P1).