Remove global sync from DG::next()
This commit removes the reduction + broadcast from DG::next(), which
essentially results in overlapping time step stages.

Now both communication steps collecting solution (unlimited then
limited) receive into their own buffers and thus are overlapped (comsol,
comlim).  After all contributions to the unlimited solution is finished
on a chare, in lim() we now combine the limited solution and the number
of degrees of freedom in cells into temporary data structures that are
created in lim() and only live until they are needed. These are ul and
ndofl in lim() and we pass them both to ownlim_complete(), which then
passes them on to dt(), where they are consumed and then discarded.
These intermediate results are not stored in DG's state (and thus they
do not increase DG's size for migration).

With this, we now only have a single synchronization point in DG during
time stepping: between dt() and solve() which is an all-to-all,
computing the minimum time step size on all chares.