Fix asynclogic problem created by removing global sync after dtref
This commit fixes the problem that was created by removing the global
barrier from [DG/DiagCG/ALECG]::resizePostAMR().

The problem was that calling Discretization::vol() without having to
wait for every chare finishing Discretization::resizePostAMR(), a chare
would not only start sending out contributions of nodal volumes to other
chares, but would also start receiving those, by receiving calls to
Discretization::comvol(). Since comvol() now could be called without
first a call to Discretization::resizePostAMR() on a chare, some data
structures that resizePostAMR() updated after mesh refinement, e.g.,
m_bid, were potentially not ready for comvol().

The solution this commit puts in still avoids the global sync by adding
a SDAG to Discretization that only continues to Transporter::vol() if
both the own and communicated contributions of the nodal volumes are
ready on a chare. In addition, comvol() now does not use m_bid
(local/boundary node id map) for storing the local nodal volume
contributions as they arrive, instead, they are stored in a map with the
global node id as the key, which is being received anyway together with
the nodal volume contributions. This makes the receive side independent
of the send side, simplifies some code, and enables correctness of this
new solution using the SDAG.