Take overlap between time steps into account
When DG::comsol() is called on a chare, the incoming array storing the
number of degrees of freedom for p-refinement, ndof, is only populated
in the 0th time step stage by the sender. However, on the receiver
chare, m_stage can still be in the previous time stage. Thus the correct
way to use the contents of ndof is if the sender's time stage was 0, not
the receiver's.

Note that execution continues on a single chare into computing and
communicating the limited solution. Even if the communication of the
unlimited solution was overlapped, yielding a stage that is unequal on
the received and the sender, this is not an issue while communicating
the limited solution, because the DAG, wait4lim(), only allows this is
both the send and the receives have been completed for the unlimited
solution, at which point the sender and receiver stage are equal. Hence
there is no need to pass the sender's stage further down.