Combine rhs and mass diffusion communication in DiagCG
There is no need to send and receive the rhs and the diffusion
contributions to the rhs in two steps as the send side is DiagCG::rhs()
for both. This simplifies code and SDAG code.

Also place the contents of DiagCG::bc() to its call side as it is only
called once. This further simplifies code in DiagCG.