Add back post-refinement global sync to ALECG/DiagCG
This commit adds back the global reduction after mesh refinement during
time stepping into ALECG and DiagCG. This appears to fix the problem
randomly occurring if this is not there. Unfortunately, I'm not sure how
to explain why this is needed based on reasoning about the code. This is
not needed in DG, but in DG there is another analogous global sync
reducing to Transporter::comfinal.

This commit also moves where the total mesh volume is stored. This is
only needed for DiagCG at this point. Previously this was stored in
DiagCG and ALECG, but now we store it in Discretization. It is a single
real number so no big deal for DG, which does not use this explicitly.
This simplifies some code around computing the total volume. In
particular, [ALECG/DiagCG/DG]::setup() now does not require an argument,
which was ignored in DG anyway.

This commit also ensures that we use the updated total volume for field
output.  Though this does not change right now, this is more correct for
future when we might distort the mesh.