Add physical-boundary leak-check on DG mesh chunk
This commit adds a verification step in the DG constructor that ensures
that the physical boundary of the input mesh partitions to all DG chares
collectively yield a non-leaky, i.e., closed, boundary.

This is done by performing a discrete surface integral over all physical
boundary faces (for which boundary conditions are assigned by the input
file), which must evaluate to zero across the whole problem if the
boundary is closed, required by a well-posed DG calculation.

This test triggers in case of a user error if a problem is configured in
a way that specifies boundary conditions (on side sets) that would yield
a problem setup with at least a single boundary cell with no boundary
condition specified.

Previously, without this leak-test in the DG constructor, this kind of
user error would trigger an assert after the initial communication step
in DG, which now provides a lot more information for debugging on which
cells are not found during communication, and thus would produce an
error message that does not explicitly point to the root of the problem.
Putting this leak-test in separates detecting the user error (not
specifying the BCs correctly for the problem) and a problem in setting
up communication data in DG.

This commit also puts in a couple of extra asserts into Discretization,
which performs mesh validity tests. Note that these tests are also done
in the Refiner constructor (which always runs, even if no mesh
refinement is configured), but running these tests once again (later) in
Discretization ensures that the mesh has not been screwed up between
Refiner and Discretization.