Replace the wrappers in Scheme with a type-matching visitor
This commit replaces all (remaining) member functions of Scheme by a
single generic member function to Scheme, which implements a
type-matching visitor on the proxy variant. As a result, there is no
longer a need for a separate SchemeBase and Scheme, and everything is
now in SchemeBase renamed to Scheme.

Client code stays simple, though the syntax changes a bit, and this
removes and significantly simplifies a lot of code. Adding a new entry
method to Scheme is now also even less work: it only requires adding a
new branch into Scheme::bcast, using 'if constexpr' as is done.