Increase tolerance for boundary integral testing leaky BCs
This commit increases the tolerance by an order of magnitude that we
still accept for the individual components of the surface integral
computed over the physical boundary involving the side sets specified by
the user for a DG calculation. This appears necessary, as this can be
triggered on some number of partitions while passing on some other
number of partitions in parallel (everything else being the same) with
data, such as

>>> Exception: Mesh boundary leaky during initialization of the DG
algorithm; this is due to incorrect or incompletely specified boundary
conditions for a given input mesh: Integral result must be a zero
vector: 6.84102397862e-14, 6.965638461e-14, 6.37719044239e-14, eps =
2.22044604925e-14

The above triggers, which it should not, because the vector is
practically zero.