AMR lib: erase tets during deref
And because now the tets are but the edges are NOT erased inside the
AMR lib, but Refiner::m_lref erases nodes belonging to erased
tets/edges, Refiner::updateEdgeData() now does not rely on querying the
AMR lib's edge_store on getting all edges in the current mesh. Instead
it queries the active_inpoel, which is up-to-date after a derefinement,
generates edges and query the refinement state of those edges. This is
to prepare sending the edge data across partitions.