Fix queries for EOS parameters
Depending on whether the EOS parameter is optional or required,
querying them from user input must be different. Note that there can be
required EOS parameters whose specification are optional for which we
then set default values (they are still required).

Examples:

  mu (viscosity) : optional, no defaults, only in viscous momentum

  k (heat conductivity): optional, no defaults, only with diffusive heat transfer

  gamma (ratio of specific heats): required, error if unspecified

  pstiff (stiffness parameter): optional, default: 0.0

  cv (specific heat at const volume): optional to specify but required, we set a default if not set

We may have to revise the above behavior in terms of user-friendliness
and more fool-proofness, but for now, because of these differences,
reading them out of g_inputdeck one must be careful, see infoCompflow().
I have put in comments for mu and k, which need not exist for the Euler eq.