Finish generation of control file keywords doc pages
This commits finishes this by removing some of the long-width (MK) links
from keywords documentations, which causes Print::splitLines() to
deadlock (this is still a problem, see #9, but for now not triggered).

Also split some keywords, e.g., diagnostics and reorder, that have been
used as both a command line as well as an input deck keyword. Since the
printout routine detect keyword aliases, and if a keyword has it, it
designates as a command line keyword, this was not easy to process when
generating the keyword reference documentation page. Conversely, these
double-keywords showed up among control file keywords too with they
alias, which was misleading. These are now split.