Various fixes to MixDir SDE
This set of changes is the result of what we have learned by trying
(initially failing) to keep a Dirichlet PDF stationary with the mixture
Dirichlet SDE.

The problem is the following: sample the initial conditions from an
arbitrarily parameterized (using omega_alpha, alpha = 1, ..., N)
Dirichlet PDF, then advance the Dirichlet SDE using constant
coefficients, b, S, and kappa, specified so that the initial PDF does
not change in time. This should be possible if Eq.(12) in
https://www.hindawi.com/journals/ijsa/2013/842981/ is satisfied.

As it turns out, one must advance the SDE in a delicate manner. In
particular, advancing the first N-1 scalars then computing the Nth one
from YN = 1 - sum_{i=1}^{N-1} Y_i may lead to numerical issues, that may
significantly alter the PDF, especially of omega_alpha << 1. A better
way to compute the Ntt scalar is to advance it via its governing
equation: dYN = - sum_{i-1}^{N-1} dYi. This commit implements this.

Another important lesson we learned is that the delicate issue of how to
treat samples that end up outside of the allowed sample space bounds,
i.e., either negative or larger than 1.0. I have tried numerous
techniques to deal with this without success, so this commit simply does
nothing if a scalar ends up out of bounds. This does not seem be a huge
problem, as the Dirichlet SDE enforces reflective boundaries at both
ends of the sample space, so if a sample (numerically) happens to leave,
it may well come back as well.

Techniques I have tried (and failed to) keeping samples within the
allowed sample space bounds were:

* Acceptance/rejection: Tried various ways of accepting updates only if
the samples stay within bounds and rejecting them otherwise, which then
re-advanced particles, but all have significantly changes the PDF.

* Detecing the negativity of the diffusion term before a time step and
doing algorithms treating this. Negative diffusion term arises from
either Yi or YN becoming out of bounds. If the sign of the whole
diffusion term is reversed, d*dW, that leads to very bad change in the
PDF. If only the coefficient d is reversed, that is similarly bad. The
best strategy to deal with negative d, to avoid negative under the
square-root is to simply zero out the diffusion coefficient, which is
what's kept for now. Also tried, re-advancing the particle with negative
diffusion term -> not good either.

This commit then introduces a new "derived" variable to the MixDir SDE:
compared to the previously stored N-1 scalars + particle density +
particle specific volume, we now also store the Nth scalar after the
(N-1)th one as this Nth scalar is now advanced by its governing
equation. This leads to having to add 1 to ncomp in the input file.
E.g., to compute ternary mixing (N = K-1 = 3), ncomp is now 5, storing
Y1, Y2, Y3=(1-Y1-Y2), rho, vol.

Thus we now never compute YN via 1.0 - sum Yi, but simply read it out
from the particle array. This actually simplifies some code in the
Dirichlet initial condition sampling as well as during advancement of
the particles.

Also added uniPDF::integrate(), which integrates a univariate PDF and
this is now also written into text-output of all univariate PDFs. This
enables verification of all univariate PDFs as valid PDFs as they should
always integrate to unity at all times to be valid PDFs.