Checkpoint/restart inciter to file appears to work
Next is to add regression tests exercising checkpoint/restart.

Details in this commit:

* Add the base class initializer to migrate constructors.

* Collect progress variable initialization to a single spot in
Transporter.hpp.  This way the enum, used for the progress variable, and
its initialization appear together, and most importantly, they can be
reused for multiple constructors.

* Abstract away info and time integration header to new member functions
Transporter::info() and Transporter::inthead(), respectively, so they
can be called at t=0 as well as returning from a restart.

* Add pup routine to single Charm++ chare, Transporter, used for
restart, and make it migratable.

* Make inciter's main chare migratable.

* Make inciter's execute chare migratable.

* Make ChareStateCollector migratable.

* Thanks to @epmikida for the help with main vs migrate constructors:
  https://lists.cs.illinois.edu/lists/arc/charm/2019-06/msg00017.html