Add two walker MixDir regression tests
These verify that we can keep a Dirichlet-sampled IC constant during
time stepping. One is keeping a double-spiked (no-mix-like) PDF, while
the other one (IJSA) is a PDF that is close to a fully mixed state, and
using the coefficients given in
https://www.hindawi.com/journals/ijsa/2013/842981. Both diff to
analytical solution, generated from evaluating the Dirichlet
distribution parameterized differently.