Send t and it from schemes instead of querying it from Transporter
4a65a239 attempted to query the physical time and iteration count from
Discretization via Scheme but that lead to random errors as, I guess,
it is not always possible to just index chare array index 0 at the time
we return from from restart.

This commit changes this by DG/ALECG,/DiagCG sending their iteration
count and physical time instead of Transporter trying to query it. This
makes Transporter::finish() and Transporter::checkpoint() a little more
complex, but it appears to work.