Move untagging of refinement and derefinement
This commit moves the removing of the refinement tags on edges marked
for refinement and the removing of the derefinement tags on edges marked
for derefinement to the ends of functions perform_refinement() and
perform_derefinement(), respectively.

With this a test case now correct that does the following initial
refinement/derefinement list: udud, where u stands for uniform
refinement, d for uniform derefinement. This used for work in serial
only, now works in parallel. New regression test for this case is next.

Credits go to @rfbird for coming up with the fix.