Comment out call to overwrite_children() from perform_refinement()
This was a work-around of the original design of AMR logic to allow
searching for the immediate parent of a tet undergoing 2:8 or 4:8. These
types of refinement steps are implemented using a backtracking step,
going from 2:1 or 4:1 followed by a 1:8, which yield 2:8 or 4:8,
respectively.

At the time the overwrite_children() was put in so that immediate
parents of nodes of 2:8 and 4:8 can be queried that yield the
intermediate tet/step, because the client algorithm in Refiner's parent
search relied on being able to query the immediate parents. This is now
done by querying only the ancestors, i.e., the parents of a node that
has no parents, i.e. going back to the very original mesh. As a result,
the AMR lib does not need to keep the intermediate state for the 2:4 and
4:8 refinement types.

This appears to not break anything with the new, ancestor-based
algorithm and passed the uncommented 'iu' regression test.

Thanks to @rfbird for figuring this out, as well.