Enable MixMassfracBeta to solve for the fluctuation
This commit enables the MixMassfractionBeta SDE to solve for the scalar
fluctuation instead of only the full scalar. This is done the same way
as for other SDEs, specifying 'solve fullvar' or 'solve fluctuation' in
their input file configuration block.

It is now an error if this is not configured within the mixmassfracbeta
... end block.

Being able to solve for the fluctuation enables to add an imposed mean
scalar gradient. This commit also fixes a previous error in how this
term was implemented (the -Gamma u_3 term needs to be multiplied by dt,
where Gamma is the imposed mean scalar gradient).

For now there is a simple specification for the mix model coefficients
in MixMassFracBetaCoeffInstVel::update(). This allows for coupled
dissipation equation but that may not be necessary for what this is
needed for. This commits adds this coupling and also cleans up the model
coefficients update.