Add compiler macro to allow turning off writing to std::cerr
By specifying the cmake variable EXCEPTIONS_WRITE_CERR as false, class
Exception now will not write to std::cerr. Even thought
src/UnitTest/QuietCerr redirects std::cerr for unittest to an internal
std::stringstream and this redirection is thread-safe most of the time,
this does not always work, perhaps due to some old libraries on our CI
machines. On newer machines this redirection appears thread safe.

This commit adds a compiler macro that can be used to turn off the
Exception class writing to std::cerr. The default is true because it
works most of the time.