Move MeshReader ctor calls to unit tests
This fixes a sneaky, and hard-to-diagnose thread-safety problem. Since
MeshReader may call ExodusIIMeshReader, which may call HDF5/NetCDF that
are not thread-safe, the MeshReader constructor, too, must be called in
a thread-safe fashion. In the unit test harness, UnitTest, this is
ensured by putting thread-unsafe calls to test groups whose name
contains the string "MPI_Single", which then the harness spawns from a
Charm++ nodegroup.

Previously to this commit, the MeshReader constructors calls have not
been executed from the thread-safe unit tests of the MeshReader group,
but from their common TUT base. This commit fixes this issue, that first
came to surface only on Mac OS X and only on Mojave.