Walker: add product variant to velocity SDE
This commit adds a capability to the velocity SDE to interpret the
dependent variable as a product to two stochastic variables. This allows
solving an SDE for particle momentum, a product of particle velocity and
fluid density. The density comes from a coupled mix model, potentially
providing multiple fluid densities.