Add more signals to trap on, enable trace
This commit adds more signals that we automatically trap on and produce
a trace for all excecutables. It also adds a commented call to the GNU
extension 'feenableexcept()' because that appears to trip on something
out of our control at this time, but would be good to enable sometime,
as it can trap NaNs and other floating point eceptions.

This commit also sets g_trace = true for all executables, which enables
a trace between program start and parsing the command line, which then
can override g_trace basde on user input. As a result, signals that
happen between program start and command line parsing always generate a
trace.