Add symmetric BC to DiagCG and system FCT
1. Symmetry BCs

This commit adds symmetry boundary conditions (BC) to the diagonal
continuous Galerkin (CG) solver. Symmetry, or free-slip, BCs can be
applied via nodes of side sets of the input mesh.

In the DiagCG constructor we compute and assign normal vectors to each
boundary node that belongs to a side set on which symmetry BCs are
specified. The normals are computed by summing the face normals, assumed
to be located in the face centers, to nodes weighted by the inverse
square distance between the face centroids and the node.

2. System FCT

This commit also adds system-character to the flux-corrected transport
(FCT).  Adding a system-character to FCT attempts to compute the limit
coefficients, for individual scalar equations that are coupled into a
system, based on the system-nature of the equations. Multiple choices
exist with different results.  For now, a variant that provides the most
monotonic (and perhaps most diffusive) results are implemented:
assigning the minimum of the limit coefficients to density, momentum,
and energy.

System-FCT is off by default. It is not necessary for independent scalar
equations. System-FCT can be enabled by the keyword 'sysfct true' in the
'inciter ... end' block.

-----------------

Besides the above two features, this commit also contains some
optimizations and code simplifications:

* DiagCG::m_bcdir (as well as the new DiagCG::m_bnorm) now assign their
data to the local (instead of global) boundary node IDs. This simplifies
code and eliminates some global->local id lookup, when BCs are
applied/verified.

* Kept skeleton functions for debugging FCT. DiagCG::writeFields() now
collects field output from FCT, to facilitate future debugging of FCT.
No fields are collected by default, but code is there to make it easy to
add both cell-, and node-centered fields for output.

* Simplified a couple of loops in FCT.

* Slightly changed how Dirichlet BCs are incorporated inside FCT. No
diffs to existing tests.

Will add Sod shock tube regression tests next.